Change order of operations in glm.nb loop
Formerly, in the iterative loop, mu (the linear predictor) was updated after updating the theta estimate

        # fit <- glm.fitter(<args>)
        th <- theta.ml(Y, mu, sum(w), w, limit=control$maxit,
                       trace = control$trace > 2)
        mu <- fit$fitted.values

This was updated to 

        # fit <- glm.fitter(<args>)
        mu <- fit$fitted.values
        th <- theta.ml(Y, mu, sum(w), w, limit=control$maxit,
                       trace = control$trace > 2)

Reason: theta parameter updates were not based off current iteration of mean model parameters, but rather the previous iteration's estimates. The change now means that theta parameter updates are based off the current iteration of mean model parameters. This change seems reasonable and is also in line with how theta is updated in the initial step.